+++ src/chrome/db/IndexDB.js
@@ -26,7 +26,9 @@ export default class IndexDB extends BaseDB {
     topList( limit = 50 ){
         return new Promise( ( resolve, reject ) => {
             let db = this.getDB();
-            resolve( db[config.dbDataTableName].orderBy('updateDate').reverse().limit( limit  ).toArray() );
+            db[config.dbDataTableName].orderBy('updateDate').reverse().limit( limit  ).toArray().then( (data)=>{
+                resolve( data );
+            });
         });
     }
 
@@ -37,7 +39,7 @@ export default class IndexDB extends BaseDB {
 		return data;
 	}
 
-    fullList( page = 1, size = 50, id, status, type = -1 ){
+    fullList( page = 1, size = 50, id, status, type = -1, searchText = '' ){
         let offset = ( page - 1 ) * size;
 
         //console.log( 'fullList', page, size, id, typeof status );
@@ -59,6 +61,12 @@ export default class IndexDB extends BaseDB {
             db[config.dbDataTableName].count(( count )=>{
                 let query = db[config.dbDataTableName].orderBy('updateDate').reverse();
 
+                if( searchText ){
+                    query.filter( ( item ) => {
+                        return item.note.indexOf( searchText ) > -1;
+                    });
+                }
+
                 if( typeof status == 'boolean' ){
 					let statusNum = status ? 1 : 0;
 					query = query.filter( ( item ) => {
@@ -74,7 +82,7 @@ export default class IndexDB extends BaseDB {
               	query.offset( offset ).limit( size ).toArray().then( ( data )=>{
 					this.fixStatus( data );
                     resolve( 
-                        { data: data, total: count }
+                        { data: data, total: data.length }
                     );
                 });
             });
@@ -90,7 +98,14 @@ export default class IndexDB extends BaseDB {
 
             db[config.dbDataTableName]
                 .orderBy('updateDate')
-                .each( ( note )=>{
+                .filter( ( item ) => {
+                    return (item.note||'').toLowerCase().indexOf( text.toLowerCase() ) > -1;
+                })
+                .toArray()
+                .then( ( data ) => {
+                    resolve( data );
+                });
+            /*.each( ( note )=>{
                     //console.log( notes );
                     let ltext = text.toLowerCase();
                     if( 
@@ -99,9 +114,11 @@ export default class IndexDB extends BaseDB {
                     ){
                         r.unshift( note );
                     }
-                }).then( ()=>{
+                })
+                .then( ( )=>{
                     resolve( r );
-                });;
+                });
+                */
 
             //resolve( [] );
         });
@@ -280,7 +297,7 @@ export default class IndexDB extends BaseDB {
     }
 
 
-    sync() {
+    sync( returnUrl ) {
         return new Promise( ( resolve, reject ) => {
             if( !this.isLogin() ){
                 resolve();
@@ -306,7 +323,7 @@ export default class IndexDB extends BaseDB {
                     console.log( 'sync', Date.now(), res );
                     this.parseRequestData( res, ()=>{
                         resolve();
-                    });
+                    }, returnUrl);
                 });
                 /*
                 this.parseRequestData( {data:sync_data}, ()=>{
@@ -358,7 +375,7 @@ export default class IndexDB extends BaseDB {
 
     }
 
-    parseRequestData( res, cb ){
+    parseRequestData( res, cb, returnUrl ){
         if( res && res.data && res.data.errno === 1 ){
             this.logout();
             return;
@@ -382,42 +399,42 @@ export default class IndexDB extends BaseDB {
                     this.batchDelete( 'md5', md5List ).then( ()=>{
                         this.batchAdd( res.data.data.sync ).then( ( data )=>{
                             this.refresh++;
-                            this.checkRefresh();
+                            this.checkRefresh(returnUrl);
                         });
                     });
                 }else{
                     this.refresh++;
-                    this.checkRefresh();
+                    this.checkRefresh(returnUrl);
                 }
                 if( res.data.data.deleted && res.data.data.deleted.length ){
                     this.batchDelete( 'md5', res.data.data.deleted ).then( ()=>{
                         this.refresh++;
-                        this.checkRefresh();
+                        this.checkRefresh(returnUrl);
                     });
                 }else{
                     this.refresh++;
-                    this.checkRefresh();
+                    this.checkRefresh(returnUrl);
                 }
 
                 console.log( 'news', res.data.data.news );
                 if( res.data.data.news && res.data.data.news.length ){
                     this.batchPush( 'md5', res.data.data.news).then( ()=>{
                         this.refresh++;
-                        this.checkRefresh();
+                        this.checkRefresh(returnUrl);
                     });
                 }else{
                     this.refresh++;
-                    this.checkRefresh();
+                    this.checkRefresh(returnUrl);
                 }
 
                 if( res.data.data.update && res.data.data.update.length ){
                     this.batchUpdate( res.data.data.update).then( ()=>{
                         this.refresh++;
-                        this.checkRefresh();
+                        this.checkRefresh(returnUrl);
                     });
                 }else{
                     this.refresh++;
-                    this.checkRefresh();
+                    this.checkRefresh(returnUrl);
                 }
 
 
@@ -427,9 +444,13 @@ export default class IndexDB extends BaseDB {
         cb && cb( res.data );
     }
 
-    checkRefresh(){
+    checkRefresh(returnUrl){
         if( this.refresh === 4 ){
-            location.reload();
+            if( returnUrl ){
+                location.replace( returnUrl );
+            }else{
+                location.reload();
+            }
         }
     }